build.py: wrapper template can be customized
default template separated to pluginwrapper.py

custom wrapper can be set in `localbuildsettings.py`, e.g.:
```
buildSettings = {
    'dev': {
        'distUrlBase': 'http://localhost:8101',
        'pluginWrapper': 'local_wrapper',  # import 'local_wrapper.py'
    },
}
```
