+++ .gitignore
@@ -1,5 +1,5 @@
 build
-localbuildsettings.py
+local*.py
 *.pyc
 *.swp
 *~
