+++ build.py
@@ -11,6 +11,7 @@ import shutil
 import json
 import shelve
 import hashlib
+from importlib import import_module  # Python >= 2.7
 
 try:
     import urllib2
@@ -59,42 +60,8 @@ distUrlBase = settings.get('distUrlBase')
 buildMobile = settings.get('buildMobile')
 gradleOptions = settings.get('gradleOptions', '')
 gradleBuildFile = settings.get('gradleBuildFile', 'mobile/build.gradle')
-
-# plugin wrapper code snippets. handled as macros, to ensure that
-# 1. indentation caused by the "function wrapper()" doesn't apply to the plugin code body
-# 2. the wrapper is formatted correctly for removal by the IITC Mobile android app
-pluginWrapperStart = """
-function wrapper(plugin_info) {
-// ensure plugin framework is there, even if iitc is not yet loaded
-if(typeof window.plugin !== 'function') window.plugin = function() {};
-
-//PLUGIN AUTHORS: writing a plugin outside of the IITC build environment? if so, delete these lines!!
-//(leaving them in place might break the 'About IITC' page or break update checks)
-plugin_info.buildName = '@@BUILDNAME@@';
-plugin_info.dateTimeVersion = '@@DATETIMEVERSION@@';
-plugin_info.pluginId = '@@PLUGINNAME@@';
-//END PLUGIN AUTHORS NOTE
-
-"""
-
-pluginWrapperStartUseStrict = pluginWrapperStart.replace("{\n", "{\n\"use strict\";\n", 1)
-
-pluginWrapperEnd = """
-setup.info = plugin_info; //add the script info data to the function as a property
-if(!window.bootPlugins) window.bootPlugins = [];
-window.bootPlugins.push(setup);
-// if IITC has already booted, immediately run the 'setup' function
-if(window.iitcLoaded && typeof setup === 'function') setup();
-} // wrapper end
-// inject code into site context
-var script = document.createElement('script');
-var info = {};
-if (typeof GM_info !== 'undefined' && GM_info && GM_info.script) info.script = { version: GM_info.script.version, name: GM_info.script.name, description: GM_info.script.description };
-script.appendChild(document.createTextNode('('+ wrapper +')('+JSON.stringify(info)+');'));
-(document.body || document.head || document.documentElement).appendChild(script);
-
-"""
-
+pluginWrapper = import_module(settings.get('pluginWrapper','pluginwrapper'))
+pluginWrapper.startUseStrict = pluginWrapper.start.replace("{\n", "{\n\"use strict\";\n", 1)
 
 pluginMetaBlock = """// @updateURL      @@UPDATEURL@@
 // @downloadURL    @@DOWNLOADURL@@
@@ -145,9 +112,9 @@ def doReplacements(script, updateUrl, downloadUrl, pluginName=None):
     script = re.sub('@@INJECTCODE@@', loadCode, script)
 
     script = script.replace('@@METAINFO@@', pluginMetaBlock)
-    script = script.replace('@@PLUGINSTART@@', pluginWrapperStart)
-    script = script.replace('@@PLUGINSTART-USE-STRICT@@', pluginWrapperStartUseStrict)
-    script = script.replace('@@PLUGINEND@@', pluginWrapperEnd)
+    script = script.replace('@@PLUGINSTART@@', pluginWrapper.start)
+    script = script.replace('@@PLUGINSTART-USE-STRICT@@', pluginWrapper.startUseStrict)
+    script = script.replace('@@PLUGINEND@@', pluginWrapper.end)
 
     script = re.sub('@@INCLUDERAW:([0-9a-zA-Z_./-]+)@@', loaderRaw, script)
     script = re.sub('@@INCLUDESTRING:([0-9a-zA-Z_./-]+)@@', loaderString, script)
